separated the stream and batch endpoints
