+++ postman.postman_collection
@@ -24,7 +24,7 @@
 							}
 						},
 						"url": {
-							"raw": "http://localhost:8080/api/v1/generate",
+							"raw": "http://localhost:8080/api/v1/generate/stream",
 							"protocol": "http",
 							"host": [
 								"localhost"
@@ -45,7 +45,7 @@
 						"method": "POST",
 						"header": [],
 						"url": {
-							"raw": "http://localhost:8080/api/v1/stop",
+							"raw": "http://localhost:8080/api/v1/generate/stream/stop",
 							"protocol": "http",
 							"host": [
 								"localhost"
@@ -80,7 +80,7 @@
 							}
 						},
 						"url": {
-							"raw": "http://localhost:8080/api/v1/generate",
+							"raw": "http://localhost:8080/api/v1/generate/batch",
 							"protocol": "http",
 							"host": [
 								"localhost"
